refactor: changed fonts to use rem instead of px
