+++ src/app/layout.tsx
@@ -1,13 +1,11 @@
 import './globals.css';
 import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
 
-import MuiThemeRegistry from '@/components/theme-registry/MuiThemeRegistry';
-
-const inter = Inter({ subsets: ['latin'] });
+import { MuiThemeRegistry } from '@/components/theme-registry/MuiThemeRegistry';
+import { PlusJakarta } from '@/font';
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
+  title: 'Pocket Network Frontend',
   description: 'Generated by create next app',
 };
 
@@ -18,7 +16,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>
+      <body className={PlusJakarta.className}>
         <MuiThemeRegistry>{children}</MuiThemeRegistry>
       </body>
     </html>
