+++ src/components/theme-registry/MuiThemeRegistry.tsx
@@ -4,14 +4,10 @@ import * as React from 'react';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider } from '@mui/material/styles';
 
-import NextAppDirEmotionCacheProvider from './EmotionCache';
-import getMuiTheme from './mui-theme';
+import { NextAppDirEmotionCacheProvider } from './EmotionCache';
+import { getMuiTheme } from './mui-theme';
 
-export default function MuiThemeRegistry({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export function MuiThemeRegistry({ children }: { children: React.ReactNode }) {
   return (
     <NextAppDirEmotionCacheProvider options={{ key: 'mui', prepend: true }}>
       <ThemeProvider theme={getMuiTheme()}>
