+++ src/font/index.ts
@@ -1,0 +1,6 @@
+import { Plus_Jakarta_Sans } from 'next/font/google';
+
+export const PlusJakarta = Plus_Jakarta_Sans({
+  subsets: ['latin'],
+  weight: ['400', '500'],
+});
